% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{.get_session}
\alias{.get_session}
\title{INTERNAL: get the session (if you still need a handle)}
\usage{
.get_session()
}
\description{
INTERNAL: get the session (if you still need a handle)
}
\keyword{internal}
