% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite-scrape.R
\name{harmonise_columns}
\alias{harmonise_columns}
\title{Make columns consistent across methods and add metadata}
\usage{
harmonise_columns(df)
}
\description{
Make columns consistent across methods and add metadata
}
\keyword{internal}
