% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite-scrape.R
\name{unite_bg_status}
\alias{unite_bg_status}
\title{Tail the background log and print pretty status}
\usage{
unite_bg_status(bg, n = 20)
}
\arguments{
\item{bg}{An object returned by \code{\link[=unite_scrape_bg]{unite_scrape_bg()}}.}

\item{n}{Number of log lines to show from the end.}
}
\description{
Tail the background log and print pretty status
}
