% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{unite_start}
\alias{unite_start}
\title{Start Selenium/Chrome (VISIBLE, global session like your script)}
\usage{
unite_start()
}
\description{
Start Selenium/Chrome (VISIBLE, global session like your script)
}
